notes/requirements.txt
dj-database-url==0.4.2
Django==2.2.3
django-ckeditor==5.3.1
django-js-asset==0.1.1
django-widget-tweaks==1.4.1
python-decouple==3.1
pytz==2017.3
sqlparse==0.3.0
